Long Awaited Update
Added:
-Plugin for Itch to make uploading easier
-Proper Soundtrack player
-Some General fixes to UI (Added a neat button graphic)

Still lots of things to do; Fixes for settings menu, parallax (which I have to code for just not added to this push yet), state-machine( I got an idea but will work with the rest of the team to make sure it works and more. Will be busy next few weeks moving and preping for Uni.